#7: Compile time import for transform * using `implicit` for `transform` implementation
